fix image

+++ app/Http/Controllers/ImageController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Log;
 
 class ImageController extends Controller
 {
@@ -10,49 +11,54 @@ class ImageController extends Controller
     {
         $header = $request->header()['authorization'][0];
         $header = base64_decode($header);
-        if($header == 'main_panel:0p3n_th!sR0ute') {
+        if ($header == 'main_panel:0p3n_th!sR0ute') {
             $url = json_decode($request->data);
             $tujuan_upload = $url->storage;
-            if($request->hasFile('file')) {
+            if ($request->hasFile('file')) {
                 $file = $request->file('file');
                 // isi dengan nama folder tempat kemana file diupload
                 $nama_file = $url->nama_file;
-                $file->move($tujuan_upload,$nama_file);
+                $file->move($tujuan_upload, $nama_file);
 
-                $fileLama = $tujuan_upload.'/'.$url->url_file;
-                if(file_exists($fileLama)) {
-                    unlink($fileLama);
-                  }
+                $fileLama = $tujuan_upload . '/' . $url->url_file;
 
+                if (file_exists($fileLama)) {
+                    if (is_file($fileLama)) {
+                        unlink($fileLama);
+                    } else {
+                        Log::warning("Path bukan file: {$fileLama}");
+                    }
+                }
             }
-            if($request->hasFile('file2')) {
+            if ($request->hasFile('file2')) {
                 $file2 = $request->file('file2');
                 // isi dengan nama folder tempat kemana file diupload
 
                 $nama_file = $url->nama_file2;
-                $file2->move($tujuan_upload,$nama_file);
+                $file2->move($tujuan_upload, $nama_file);
 
-                $fileLama = $tujuan_upload.'/'.$url->url_file2;
-                if(file_exists($fileLama)) {
-                    unlink($fileLama);
-                  }
+                $fileLama = $tujuan_upload . '/' . $url->url_file2;
 
+                if (file_exists($fileLama)) {
+                    if (is_file($fileLama)) {
+                        unlink($fileLama);
+                    } else {
+                        Log::warning("Path bukan file: {$fileLama}");
+                    }
+                }
             }
             return [
                 'code'      => 200,
-    			'response' => true,
-    			'message' => 'success'
-    		];
+                'response' => true,
+                'message' => 'success'
+            ];
         } else {
             return [
                 'code'      => 200,
-    			'response' => false,
-    			'message' => 'authorization is fail!'
-    		];
+                'response' => false,
+                'message' => 'authorization is fail!'
+            ];
         }
-
-
-
     }
 
     public function hapusQr(Request $request)
@@ -60,25 +66,21 @@ class ImageController extends Controller
         $header = $request->header()['authorization'][0];
         $header = base64_decode($header);
         $fileLama = $request->nama_file;
-        if($header == 'main_panel:0p3n_th!sR0ute') {
-            if(file_exists($fileLama)) {
+        if ($header == 'main_panel:0p3n_th!sR0ute') {
+            if (file_exists($fileLama)) {
                 unlink($fileLama);
-              }
+            }
             return [
                 'code'      => 200,
-    			'response' => true,
-    			'message' => 'success'
-    		];
-        }else {
+                'response' => true,
+                'message' => 'success'
+            ];
+        } else {
             return [
                 'code'      => 200,
-    			'response' => false,
-    			'message' => 'authorization is fail!'
-    		];
+                'response' => false,
+                'message' => 'authorization is fail!'
+            ];
         }
-
-
-
     }
-
 }
